Chore: rename factory module
