+++ docs/current/careamist_predicting.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 from pathlib import Path
 import numpy as np
-from careamics.config.ng_factories import create_n2v_config, create_care_config
+from careamics.config.factories import create_n2v_config, create_care_config
 from careamics.dataset_ng.factory import ReadFuncLoading
 from careamics.careamist_v2 import CAREamistV2
 import tifffile
