+++ docs/current/configuration_io.py
@@ -2,10 +2,10 @@
 
 # %%
 # --8<-- [start:save_load]
-from careamics.config.ng_factories import create_n2v_config
+from careamics.config.factories import create_n2v_config
 from careamics.config.utils.configuration_io import (
     save_configuration,
-    load_configuration_ng,
+    load_configuration,
 )
 
 # create a configuration
@@ -22,5 +22,5 @@ config = create_n2v_config(
 config_path = save_configuration(config, "careamics_config.yml")
 
 # load the configuration
-loaded_config = load_configuration_ng(config_path)
+loaded_config = load_configuration(config_path)
 # --8<-- [end:config]
