+++ docs/current/configuration_simple.py
@@ -2,7 +2,7 @@
 
 # %%
 # --8<-- [start:all_configs]
-from careamics.config.ng_factories import (
+from careamics.config.factories import (
     create_n2v_config,
     create_care_config,
     create_n2n_config,
@@ -17,7 +17,7 @@ from careamics.config.ng_factories import (
 ################ Noise2Void ################
 # %%
 # --8<-- [start:config_n2v]
-from careamics.config.ng_factories import create_n2v_config
+from careamics.config.factories import create_n2v_config
 
 # create a configuration
 config = create_n2v_config(
@@ -32,7 +32,7 @@ config = create_n2v_config(
 
 # %%
 # --8<-- [start:config_n2v_steps]
-from careamics.config.ng_factories import create_n2v_config
+from careamics.config.factories import create_n2v_config
 
 # create a configuration
 config = create_n2v_config(
@@ -48,7 +48,7 @@ config = create_n2v_config(
 
 # %%
 # --8<-- [start:config_n2v_augs]
-from careamics.config.ng_factories import create_n2v_config
+from careamics.config.factories import create_n2v_config
 
 # create a configuration
 config = create_n2v_config(
@@ -64,7 +64,7 @@ config = create_n2v_config(
 
 # %%
 # --8<-- [start:config_n2v_channels]
-from careamics.config.ng_factories import create_n2v_config
+from careamics.config.factories import create_n2v_config
 
 # create a configuration
 config = create_n2v_config(
@@ -80,7 +80,7 @@ config = create_n2v_config(
 
 # %%
 # --8<-- [start:config_n2v_val]
-from careamics.config.ng_factories import create_n2v_config
+from careamics.config.factories import create_n2v_config
 
 # create a configuration
 config = create_n2v_config(
@@ -99,7 +99,7 @@ config = create_n2v_config(
 ################# CARE/N2N #################
 # %%
 # --8<-- [start:config_care]
-from careamics.config.ng_factories import create_care_config
+from careamics.config.factories import create_care_config
 
 # create a configuration
 config = create_care_config(
@@ -114,7 +114,7 @@ config = create_care_config(
 
 # %%
 # --8<-- [start:config_n2n]
-from careamics.config.ng_factories import create_n2n_config
+from careamics.config.factories import create_n2n_config
 
 # create a configuration
 config = create_n2n_config(
@@ -129,7 +129,7 @@ config = create_n2n_config(
 
 # %%
 # --8<-- [start:config_care_steps]
-from careamics.config.ng_factories import create_care_config
+from careamics.config.factories import create_care_config
 
 # create a configuration
 config = create_care_config(
@@ -145,7 +145,7 @@ config = create_care_config(
 
 # %%
 # --8<-- [start:config_care_augs]
-from careamics.config.ng_factories import create_care_config
+from careamics.config.factories import create_care_config
 
 # create a configuration
 config = create_care_config(
@@ -161,7 +161,7 @@ config = create_care_config(
 
 # %%
 # --8<-- [start:config_care_channels]
-from careamics.config.ng_factories import create_care_config
+from careamics.config.factories import create_care_config
 
 # create a configuration
 config = create_care_config(
@@ -178,7 +178,7 @@ config = create_care_config(
 
 # %%
 # --8<-- [start:config_care_val]
-from careamics.config.ng_factories import create_care_config
+from careamics.config.factories import create_care_config
 
 # create a configuration
 config = create_care_config(
