+++ docs/current/getting_started.py
@@ -12,7 +12,7 @@ pred_data = train_data
 # %%
 # --8<-- [start:quick_start_n2v]
 from careamics.careamist_v2 import CAREamistV2
-from careamics.config.ng_factories import create_n2v_config
+from careamics.config.factories import create_n2v_config
 
 # create a configuration
 config = create_n2v_config(
@@ -36,7 +36,7 @@ prediction = careamist.predict(pred_data=pred_data)
 
 # --8<-- [start:quick_start_care]
 from careamics.careamist_v2 import CAREamistV2
-from careamics.config.ng_factories import create_care_config
+from careamics.config.factories import create_care_config
 
 # create a configuration
 config = create_care_config(
