+++ docs/current/training_logging.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import numpy as np
-from careamics.config.ng_factories import create_n2v_config, create_care_config
+from careamics.config.factories import create_n2v_config
 from careamics.careamist_v2 import CAREamistV2
 
 train_data = np.random.randint(0, 255, (512, 512)).astype(np.float32)
