+++ src/careamics/config/factories/algorithm_factory.py
