+++ src/careamics/config/factories/care_n2n_factory.py
