+++ src/careamics/config/factories/data_factory.py
