+++ src/careamics/config/factories/training_factory.py
