+++ src/careamics/config/support/supported_transforms.py
@@ -8,5 +8,4 @@ class SupportedTransform(str, BaseEnum):
 
     XY_FLIP = "XYFlip"
     XY_RANDOM_ROTATE90 = "XYRandomRotate90"
-    NORMALIZE = "Normalize"
     N2V_MANIPULATE = "N2VManipulate"
