+++ src/careamics/lightning/dataset_ng/lightning_modules/care_module.py
@@ -9,7 +9,7 @@ from torch import nn
 from torchmetrics import MetricCollection
 
 from careamics.config import CAREAlgorithm, N2NAlgorithm
-from careamics.config.ng_factories.algorithm_factory import algorithm_factory
+from careamics.config.factories.algorithm_factory import algorithm_factory
 from careamics.config.support import SupportedLoss
 from careamics.dataset_ng.dataset import ImageRegionData
 from careamics.lightning.dataset_ng.data_module import TrainValData, TrainValSplitData
