+++ tests/interoperability/test_lightning_module_onnx_exportability.py
@@ -2,7 +2,7 @@ import pytest
 import torch
 from onnx import checker
 
-from careamics.config.ng_factories.algorithm_factory import algorithm_factory
+from careamics.config.factories.algorithm_factory import algorithm_factory
 from careamics.lightning.dataset_ng.lightning_modules import get_module_cls
