+++ tests/test_smoke_after_packaging.py
@@ -5,7 +5,7 @@
 def test_smoke():
     """Import CAREamist."""
     from careamics import CAREamistV2
-    from careamics.config.ng_factories import create_n2v_config
+    from careamics.config.factories import create_n2v_config
 
     # create configuration
     config = create_n2v_config(
